requirements.txt
numpy
pandas
matplotlib
torch
torchvision
spacy
flask
requests
bs4
pyttsx3
google-cloud-texttospeech